Fixed path to admin's selector search image
With 1.3's django.contrib.staticfiles, after a run of `collect static`
static assets are copied to the more conventional location of `{{
STATIC_URL }}admin/img/selector-search.gif`. This patch fixes it for
now, but the ADMIN_MEDIA_URL prefix is deprecated in the upcoming 1.4.
This should be rewritten with STATIC_URL soon.